% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientPickerD3.R
\name{gradientPickerD3-shiny}
\alias{gradientPickerD3-shiny}
\alias{gradientPickerD3Output}
\alias{renderGradientPickerD3}
\title{Shiny bindings for gradientPickerD3}
\usage{
gradientPickerD3Output(outputId, width = "100\%", height = "100\%")

renderGradientPickerD3(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a gradientPickerD3}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using gradientPickerD3 within Shiny
applications and interactive Rmd documents.
}
