% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientPickerD3_example.R
\name{gradientPickerD3_example}
\alias{gradientPickerD3_example}
\title{gradientPickerD3_example}
\usage{
gradientPickerD3_example()
}
\description{
Creates an example shiny app which include the gradientPickerD3 and a rendered table for gradientPickerD3 return value. By clicking the reload button new random ticks will be generated.
}
