% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientPickerD3.R
\name{gradientPickerD3}
\alias{gradientPickerD3}
\title{gradientPickerD3}
\usage{
gradientPickerD3(payload, width = NULL, height = NULL, elementId = NULL,
  border_extensions = 0.001)
}
\arguments{
\item{payload}{list containing 'ticks' and 'colors' to init the gradient}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},}

\item{elementId}{output variable to read from}

\item{border_extensions}{add to the min and max data range to cover the whole color spectrum}
}
\description{
creates an interactive color gradient for shiny
}
